spec cleanup
