+++ spec/csl/styles/tex_citeauthors_numeric_spec.rb
@@ -3,7 +3,7 @@
 require_relative "../../citation_helper"
 
 describe "citeauthor* macro with numeric style" do
-  let(:options) { { "bibliography-tex-style" => "numeric" } }
+  let(:options) { { "bibliography-tex-style" => "numeric", "bibliography-database" => "database.bib" } }
 
   it "formats a single citation" do
     expect(formatted_citation("citeauthor*:[Erdos65]", options: options)).
