+++ spec/fixtures/database.bib
@@ -29,3 +29,33 @@
 	pages = {1036--1060},
 	year = {2004}
 }
+
+@article{Erdos65,
+  title = {Some very hard sums},
+  journal={Difficult Maths Today},
+  author={Paul Erdős and Arend Heyting and Luitzen Egbertus Brouwer},
+  year={1965},
+  pages={30}
+}
+
+@article{Einstein35,
+  title={Can quantum-mechanical description of physical reality be considered complete?},
+  author={Einstein, Albert and Podolsky, Boris and Rosen, Nathan},
+  journal={Physical review},
+  volume={47},
+  number={10},
+  pages={777},
+  year={1935},
+  publisher={APS}
+}
+
+@article{Gettier63,
+  title={Is justified true belief knowledge?},
+  author={Gettier, Edmund L},
+  journal={analysis},
+  volume={23},
+  number={6},
+  pages={121--123},
+  year={1963},
+  publisher={JSTOR}
+}
